Remove functionalities from main.rb and create new file for music methods
