+++ books.json
@@ -1,1 +1,1 @@
-[{"id":6,"type":"Book","publish_date":"2021-12-21","archived":false,"title":"book1","publisher":"pub1","cover_state":"good"},{"id":7,"type":"Book","publish_date":"2000-03-21","archived":false,"title":"book2","publisher":"pub2","cover_state":"good"},{"id":8,"type":"Book","publish_date":"2012-04-06","archived":false,"title":"book4","publisher":"pub4","cover_state":"good"},{"id":9,"type":"Book","publish_date":"2012-07-12","archived":false,"title":"book5","publisher":"pub5","cover_state":"good"},{"id":10,"type":"Book","publish_date":"2020-02-02","archived":false,"title":"tit1","publisher":"pub2","cover_state":"good"}]
+[{"id":7,"type":"Book","publish_date":"2021-12-21","archived":false,"title":"book1","publisher":"pub1","cover_state":"good"},{"id":8,"type":"Book","publish_date":"2000-03-21","archived":false,"title":"book2","publisher":"pub2","cover_state":"good"},{"id":9,"type":"Book","publish_date":"2012-04-06","archived":false,"title":"book4","publisher":"pub4","cover_state":"good"},{"id":10,"type":"Book","publish_date":"2012-07-12","archived":false,"title":"book5","publisher":"pub5","cover_state":"good"},{"id":11,"type":"Book","publish_date":"2020-02-02","archived":false,"title":"tit1","publisher":"pub2","cover_state":"good"},{"id":12,"type":"Book","publish_date":"2021-09-09","archived":false,"title":"book6","publisher":"pub6","cover_state":"good"}]
