+++ json_files/genres.json
@@ -1,1 +1,1 @@
-[{"name":"Ode"},{"name":"Country music"},{"name":"pop"},{"name":"gen1"},{"name":"gen"}]
+[{"name":"Ode"},{"name":"Country music"},{"name":"pop"},{"name":"gen1"},{"name":"gen"},{"name":"gen29"}]
