+++ json_files/music_album.json
@@ -1,1 +1,1 @@
-[{"name":"Ode","publish_date":"2002-01-01","on_spotify":true},{"name":"Neolicious","publish_date":"1991-02-05","on_spotify":true},{"name":"thrill of it all","publish_date":"2020-08-05","on_spotify":true},{"name":"alb1","publish_date":"2021-09-21","on_spotify":true},{"name":"alb","publish_date":"2021-09-21","on_spotify":true}]
+[{"name":"Ode","publish_date":"2002-01-01","on_spotify":true},{"name":"Neolicious","publish_date":"1991-02-05","on_spotify":true},{"name":"thrill of it all","publish_date":"2020-08-05","on_spotify":true},{"name":"alb1","publish_date":"2021-09-21","on_spotify":true},{"name":"alb","publish_date":"2021-09-21","on_spotify":true},{"name":"alb29","publish_date":"2002-09-09","on_spotify":true}]
